Add MISRA-C:2012 compliance analysis and gate
Implements TODO item 4: a MISRA-C:2012 compliance check via cppcheck's misra
addon, wired into the build as a gating analyzer with a documented set of
justified deviations. Establishing the analysis also surfaced and fixed 12
genuine type-safety issues in the firmware.

Tooling (make analyze-misra)
----------------------------
New `analyze-misra` target runs the cppcheck misra addon over every firmware
translation unit (core + each output-driver variant), each under a
representative variant -D. It GATES: any finding not covered by a documented
deviation fails the build. It is part of the `analyze` aggregate and therefore
of `make test`. A companion `analyze-misra-report` prints the full inventory
including the waived deviations (for maintaining the compliance doc).

Supporting files:
  - test/misra.json            addon config (points misra.py at the rule texts)
  - test/misra_rules.txt       our own short paraphrases of each rule (the
                               official MISRA texts are copyrighted and not
                               shipped by cppcheck; without them every finding
                               is an opaque number)
  - test/misra_suppressions.txt  per-file deviation waivers, each mapped to a
                               record in MISRA_COMPLIANCE.md

Notes captured in the Makefile/comments for reproducibility: PYTHONWARNINGS=
ignore is required (cppcheck treats the addon's Py3.12+ DeprecationWarning as a
hard failure); avr-libc/avr-gcc system headers are excluded as outside the
compliance boundary; and a MISRA-specific avr-libc include discovery is used
because `avr-gcc -print-file-name=avr/io.h` returns a bare name on this
toolchain (the value rules are meaningless without the real register headers).

Firmware type-safety fixes (12 findings cleared)
------------------------------------------------
These are genuine improvements, not tool appeasement:
  - PRESSED_THRESH / RELEASE_THRESH are now unsigned (a debounce count is
    non-negative), clearing essential-type mismatches at the comparison sites.
  - Unsigned literals in the saturating-integrator and sanity-check comparisons
    (0U/1U), and an unsigned timer-formula assert.
  - New DEBOUNCE_COUNTER_MAX (255U) replaces UINT8_MAX in the threshold bound
    asserts: UINT8_MAX is essentially SIGNED (C integer promotion expands it to
    (0x7f*2+1)), so comparing it to the unsigned thresholds tripped rules 10.4
    and 12.1. The unsigned constant means the same thing and clears both. It is
    defined in bypass_core.c (its only user) to avoid a cross-TU "unused macro".

All variants build clean across both clocks under -Wall -Wextra -Werror
-Wconversion; behavior is unchanged (type-only edits).

Documented deviations (MISRA_COMPLIANCE.md)
-------------------------------------------
The remaining 63 findings are two well-justified classes, each waived per-file
(not project-wide), so a new occurrence in a new file still fails the gate:
  - D-1  Rules 11.4 / 10.1 / 10.8 - hardware register access via avr-libc's
         _SFR_* macros (integer-to-pointer conversion, plus bit ops and
         composite casts in register read-modify-write). Unavoidable on
         bare-metal AVR.
  - D-2  Rule 2.5 - FOOTSW_PIN / LED_PIN defined once in the shared
         bypass_output_common.h, reported "unused" when cppcheck analyzes a
         driver TU that does not reference them. Not dead code.

Test infrastructure
-------------------
Making the thresholds unsigned rippled into two host model/proof files that do
signed state-index and loop arithmetic under -Werror=sign-*. Cast RELEASE_THRESH
to int at those sites (test_model_check.c: COUNTER_VALUES macro + two loops;
test_symbolic.c: one loop). No logic change.

Verification
------------
  - make test: green (all variants x MCUs, fault-injection, 100% model coverage)
  - make analyze-misra: clean (63 deviations waived per MISRA_COMPLIANCE.md)
  - gate confirmed to fail on a new, un-waived finding (injected 10.4 -> exit 2)

Files added: MISRA_COMPLIANCE.md, test/misra.json, test/misra_rules.txt,
test/misra_suppressions.txt. .gitignore now ignores cppcheck addon artifacts
(*.dump, *.ctu-info, cppcheck-addon-ctu-file-list*).

+++ test/test_model_check.c
@@ -69,7 +69,10 @@ static int g_checks = 0;
 } while (0)
 
 // Dense index for a state. Counter range is 0..RELEASE_THRESH inclusive.
-#define COUNTER_VALUES (RELEASE_THRESH + 1)
+// RELEASE_THRESH is unsigned in the firmware config (a count is non-negative);
+// cast to int here so this file's signed index/loop arithmetic stays signed and
+// does not trip -Werror=sign-conversion/sign-compare.
+#define COUNTER_VALUES ((int)RELEASE_THRESH + 1)
 #define NUM_STATES     (2 * 2 * COUNTER_VALUES)
 
 static int state_index(state_t s) {
@@ -167,7 +170,7 @@ static void verify_lockout_requires_full_release(void) {
     // From any RELEASE_DEBOUNCE_WAIT state, the ONLY transition back to
     // PRESS_DEBOUNCE_WAIT requires debounce_counter == 0 after the ISR update.
     for (uint8_t es = 0; es <= ENGAGED; ++es) {
-        for (int dc = 0; dc <= RELEASE_THRESH; ++dc) {
+        for (int dc = 0; dc <= (int)RELEASE_THRESH; ++dc) {
             state_t s = { RELEASE_DEBOUNCE_WAIT, es, (uint8_t)dc };
             for (int bit = 0; bit < 2; ++bit) {
                 step_result_t r = step(s, bit);
@@ -193,7 +196,7 @@ static void verify_release_liveness(void) {
 
     for (uint8_t ps = 0; ps <= RELEASE_DEBOUNCE_WAIT; ++ps) {
         for (uint8_t es = 0; es <= ENGAGED; ++es) {
-            for (int dc = 0; dc <= RELEASE_THRESH; ++dc) {
+            for (int dc = 0; dc <= (int)RELEASE_THRESH; ++dc) {
                 state_t s = { ps, es, (uint8_t)dc };
                 int reached = 0;
                 int toggles = 0;
